refactor: add RIF rebrand REM-29 (#540)
* refactor: rif rebrand

* feat: adding qa deployment workflow

* fix: node version

* fix: add mainnet addresses

* fix: set qa workflow to build mainnet enviroment

* fix: make qa workflow to deploy once pushing to disposable-main branch

* fix: make styles match testnet version

* fix: replacing alt from RSK to RIF logo

* fix: remove old font

* fix: keeping  usage

* fix: keeping  usage

* fix: keeping  usage

* fix: adding 26853custom-gray

* fix: setting string resolver to address zero

* refactor: login form rebrand
