+++ .github/workflows/qa-deploy.yml
@@ -1,0 +1,36 @@
+name: RNS QA build and deploy
+
+on:
+  push:
+    branches: [ disposable-main ]
+
+jobs:
+  qa_build_and_deploy:
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Setup Node 12
+        uses: actions/setup-node@v1
+        with:
+          node-version: '12'
+
+      - name: Install dependencies
+        run: yarn
+
+      - name: Build site
+        run: yarn build:mainnet
+
+      - name: Configure AWS credentials
+        uses: aws-actions/configure-aws-credentials@v1
+        with:
+          aws-access-key-id: ${{ secrets.QA_AWS_ACCESS_KEY_ID }}
+          aws-secret-access-key: ${{ secrets.QA_AWS_SECRET_ACCESS_KEY }}
+          aws-region: ${{ secrets.TESTNET_AWS_REGION }}
+
+      - name: Deploy site to S3
+        run: |
+          aws s3 sync --delete --only-show-errors build/ ${{ secrets.QA_S3_BUCKET }}
