+++ src/assets/css/OverwriteBootstrap.css
@@ -22,7 +22,7 @@ h3,
 h4,
 h5,
 h6 {
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 300;
 }
 
@@ -40,7 +40,7 @@ h2 b,
 h3 b,
 h4 b {
     font-weight: 700;
-    color: #008CFF;
+    color: #4B5CF0;
 }
 
 h5 strong,
@@ -80,7 +80,7 @@ h6 {
 }
 
 p {
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     color: black;
     font-size: 16px;
 }
@@ -92,7 +92,7 @@ p {
 
 .custom-dot >li::before {
     content: "\2022";
-    color: #008CFF;
+    color: #4B5CF0;
     font-weight: bold;
     display: inline-block;
     width: 12px;
@@ -119,7 +119,7 @@ p {
 
 
 a {
-    color: #008CFF;
+    color: #4B5CF0;
     text-decoration: none;
 }
 
@@ -136,7 +136,7 @@ a:hover {
 /*  ALERT  */
 
 .alert {
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 400;
     font-size: 14px;
     word-break: break-all;
@@ -210,20 +210,20 @@ a:hover {
 
 .btn {
     border-radius: 40px;
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-size: 14px;
     padding: 3px 15px!important;
 }
 
 .btn-primary {
-    color: #FFFFFF;
-    background-color: #008CFF;
-    border-color: #008CFF;
+    color: #BFEEFF;
+    background-color: #4B5CF0;
+    border-color: #4B5CF0;
     font-weight: 500;
 }
 
 .btn-secondary {
-    color: #008CFF;
+    color: #4B5CF0;
     background-color: #FFFFFF;
     border-color: #FFFFFF;
     font-weight: 500;
@@ -237,16 +237,16 @@ a:hover {
 }
 
 .btn-info {
-    color: #008CFF;
+    color: #4B5CF0;
     background-color: #FFFFFF;
-    border-color: #008CFF;
+    border-color: #4B5CF0;
     font-weight: 400;
 }
 
 .btn-light {
     color: #FFFFFF;
-    background-color: #008CFF;
-    border-color: #008CFF;
+    background-color: #4B5CF0;
+    border-color: #4B5CF0;
     font-weight: 500;
 }
 
@@ -257,7 +257,7 @@ a:hover {
 }
 
 .btn-primary:hover:not([disabled]) {
-    color: #008CFF;
+    color: #4B5CF0;
     background-color: #FFFFFF;
     border-color: #FFFFFF;
     font-weight: 700;
@@ -266,8 +266,8 @@ a:hover {
 
 .btn-secondary:hover {
     color: #FFFFFF;
-    background-color: #008CFF;
-    border-color: #008CFF;
+    background-color: #4B5CF0;
+    border-color: #4B5CF0;
     font-weight: 700;
     box-shadow: 5px 5px 5px #333333;
 }
@@ -282,13 +282,13 @@ a:hover {
 
 .btn-info:hover {
     color: #FFFFFF;
-    background-color: #008CFF;
+    background-color: #4B5CF0;
     border-color: #FFFFFF;
     font-weight: 400;
 }
 
 .btn-light:hover {
-    color: #008CFF;
+    color: #4B5CF0;
     background-color: #FFFFFF;
     border-color: #FFFFFF;
     font-weight: 500;
@@ -349,9 +349,9 @@ a:hover {
 /*  MODAL   */
 
 .modal-header {
-    background-color: #008CFF;
+    background-color: #4B5CF0;
     color: #FFFFFF;
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     padding: 0.50rem 30px;
 }
 
@@ -388,12 +388,12 @@ a:hover {
 
 .toast-header {
     color: black;
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 700;
 }
 
 .toast-body {
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 300;
     text-align: left;
 }
@@ -402,18 +402,18 @@ a:hover {
 /*  POPOVER */
 
 .popover {
-    border-color: #008CFF;
+    border-color: #4B5CF0;
 }
 
 .popover-header {
     background-color: white;
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 500;
 }
 
 .popover-body {
     background-color: white;
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 300;
 }
 
@@ -422,7 +422,7 @@ a:hover {
 }
 
 .popover .arrow {
-    color: #008CFF;
+    color: #4B5CF0;
 }
 
 
@@ -433,7 +433,7 @@ a:hover {
     font-size: 14px;
     background: transparent;
     font-weight: 300;
-    border: solid 1px #008CFF;
+    border: solid 1px #4B5CF0;
 }
 
 .form-control::-webkit-input-placeholder {
@@ -489,7 +489,7 @@ a:hover {
 }
 
 .input-group-text {
-    color: #008CFF;
+    color: #4B5CF0;
     background-color: transparent;
     border: none;
     border-radius: 0rem;
