+++ src/assets/css/ReplaceBootstrap.css
@@ -2,7 +2,7 @@
 
 /*  TEXT    */
 h1, h2, h3, h4, h5 {
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 300;
 }
 
@@ -12,16 +12,16 @@ h5 strong, h6 strong {
 
 h1 strong, h2 strong, h3 strong {
     font-weight: 700;
-    color: #008CFF;
+    color: #4B5CF0;
 }
 
 p {
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     color: black;
 }
 
 ul li {
-    color: #008CFF;
+    color: #4B5CF0;
 }
 
 ul li ul li {
@@ -29,7 +29,7 @@ ul li ul li {
  }
 
 ol li {
-    color: #008CFF;
+    color: #4B5CF0;
 }
 
 ol li ul li {
@@ -41,7 +41,7 @@ ol li ul li {
 
 /*  ALERT  */
 .alert {
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 500;
 }
 .alert-primary-rif {
@@ -103,7 +103,7 @@ ol li ul li {
 /*  BUTTON  */
 .btn {
     border-radius: 40px;
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-size:14px;
     padding: 3px 15px;
 }
@@ -115,13 +115,13 @@ ol li ul li {
 
 .btn-primary-rif {
     color: #FFFFFF;
-    background-color: #008CFF;
-    border-color: #008CFF;
+    background-color: #4B5CF0;
+    border-color: #4B5CF0;
     font-weight: 700;
 }
 
 .btn-secondary-rif {
-    color: #008CFF;
+    color: #4B5CF0;
     background-color: #FFFFFF;
     border-color: #FFFFFF;
     font-weight: 700;
@@ -135,21 +135,21 @@ ol li ul li {
 }
 
 .btn-info-rif {
-    color: #008CFF;
+    color: #4B5CF0;
     background-color: #FFFFFF;
-    border-color: #008CFF;
+    border-color: #4B5CF0;
     font-weight: 400;
 }
 
 .btn-light-rif {
     color: #FFFFFF;
-    background-color: #008CFF;
-    border-color: #008CFF;
+    background-color: #4B5CF0;
+    border-color: #4B5CF0;
     font-weight: 500;
 }
 
 .btn-primary-rif:hover {
-    color: #008CFF;
+    color: #4B5CF0;
     background-color: #FFFFFF;
     border-color: #FFFFFF;
     font-weight: 700;
@@ -158,8 +158,8 @@ ol li ul li {
 
 .btn-secondary-rif:hover {
     color: #FFFFFF;
-    background-color: #008CFF;
-    border-color: #008CFF;
+    background-color: #4B5CF0;
+    border-color: #4B5CF0;
     font-weight: 700;
     box-shadow: 5px 5px 5px #333333;
 }
@@ -174,13 +174,13 @@ ol li ul li {
 
 .btn-info-rif:hover {
     color: #FFFFFF;
-    background-color: #008CFF;
+    background-color: #4B5CF0;
     border-color: #FFFFFF;
     font-weight: 400;
 }
 
 .btn-light-rif:hover {
-    color: #008CFF;
+    color: #4B5CF0;
     background-color: #FFFFFF;
     border-color: #FFFFFF;
     font-weight: 500;
@@ -237,9 +237,9 @@ ol li ul li {
 
 /*  MODAL   */
 .modal-header-rif {
-    background-color: #008CFF;
+    background-color: #4B5CF0;
     color: #FFFFFF;
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
 }
 .modal-title-rif{font-weight:700;}
 
@@ -267,32 +267,32 @@ ol li ul li {
 /*  TOAST   */
 .toast-header-rif {
     color: black;
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 700;
 }
 
 .toast-body-rif {
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 300;
 }
 
 
 /*  POPOVER */
 .popover-rif{
-    border-color: #008CFF;
+    border-color: #4B5CF0;
 }
 
 
 
 .popover-header-rif {
     background-color: white;
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 500;
 }
 
 .popover-body-rif {
     background-color: white;
-    font-family: 'Rubik', sans-serif;
+    font-family: 'Sora', sans-serif;
     font-weight: 300;
 }
 
@@ -301,12 +301,12 @@ ol li ul li {
 }
 
 .popover-rif .arrow {
-    color: #008CFF;
+    color: #4B5CF0;
 }
 
 
 /* INPUT */
-.form-control-rif{color:#000000; font-size:14px; background:transparent;font-weight: 300; border:solid 1px #008CFF; }
+.form-control-rif{color:#000000; font-size:14px; background:transparent;font-weight: 300; border:solid 1px #4B5CF0; }
 .form-control-rif::-webkit-input-placeholder { /* Edge */
   color:#000000; font-weight: 300;
 }
