+++ src/assets/css/sass/_index.scss
@@ -1,6 +1,7 @@
 $gray: #919191;
 $light-gray: #f7f7f7;
-$blue: #008CFF;
+$custom-gray: #E6DED6;
+$blue: #4B5CF0;
 $white: #ffffff;
 $red: #E02F2F;
 $light-red: rgba(232, 99, 99, 0.2);
