+++ src/assets/css/sass/admin/_index.scss
@@ -7,7 +7,7 @@
   @import 'resolver';
 
   margin-top: 75px;
-  
+
   .btn-primary:not(:disabled):not(.disabled).active {
     background-color: $light-gray;
     border-color: $gray;
