+++ src/assets/css/sass/admin/_subdomains.scss
@@ -1,7 +1,7 @@
 .subdomains {
   .owner, .subdomain {
     width: 100%;
-    padding:10px 10px; 
+    padding:10px 10px;
     border-radius: 12px;
     border: 1px solid $gray;
   }
