+++ src/assets/css/sass/components/_addressInputComponent.scss
@@ -72,7 +72,7 @@
       border-radius: 15px;
       width: 47%
     }
-    
+
     button.cancel {
       border:1px solid $gray;
       background: $white;
