+++ src/assets/css/sass/components/_loginDropdownContainer.scss
@@ -6,7 +6,7 @@
     width: 300px;
     right: 0px;
     top: 50px;
-    background-color: $white;
+    background-color: $custom-gray;
     box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2);
     border-radius: 12px;
     list-style-type: none;
@@ -23,7 +23,7 @@
         display: block;
         width: 100%;
         border: none;
-        background-color: #fff;
+        background-color: $custom-gray;
         font-size: 14px;
         text-align: left;
       }
@@ -67,7 +67,7 @@
 
     .loginForm {
       padding: 10px 15px;
-      
+
       button.active {
         color: $blue;
       }
@@ -75,9 +75,10 @@
       h3 {
         font-size: 14px;
       }
-      
+
       .rskinput {
-        border: 1px solid #E5E5E5;
+        background-color: $white;
+        border: 1px solid #919191;
         border-radius: 5px;
         padding: 5px;
 
@@ -96,8 +97,8 @@
 
       .btn.btn-primary {
         margin: 10px auto 0 auto !important;
-        background-color: #C4C4C4;
-        border-color: #C4C4C4;
+        background-color: #919191;
+        border-color: #919191;
         border-radius: 15px;
         width: auto;
       }
