+++ src/assets/css/sass/components/_share.scss
@@ -17,7 +17,7 @@ button.share-button:hover {
     h2 {
       font-size: 16px !important;
     }
-    
+
     .row {
       min-width: 250px;
