+++ src/assets/css/sass/components/_suggestionsList.scss
