+++ src/assets/css/sass/components/_toggleContainer.scss
@@ -54,5 +54,5 @@
 
   button.active {
     color: $blue;
-  }  
+  }
 }
