+++ src/assets/css/sass/components/_userDeleteComponent.scss
@@ -10,7 +10,7 @@ div.delete {
     width: 45% !important;
     border: 1px solid $gray !important;
   }
-  
+
   button.btn-primary.cancel {
     background: $white;
     color: $gray;
