+++ src/assets/css/sass/home/_index.scss
@@ -15,7 +15,7 @@
     }
 
     h1, p {
-      color: #ffffff;
+      color: $custom-gray;
       text-align: center;
     }
 
@@ -35,7 +35,7 @@
     content: '';
     width: 100%;
     height: 80px;
-    background: url("/assets/img/home/bottom-background.svg");
+    background: $blue;
     background-position: center bottom;
     background-repeat: no-repeat;
     background-size: cover;
@@ -94,7 +94,7 @@
       }
 
       .carousel-inner {
-        a { 
+        a {
           display: block;
 
           p {
@@ -130,7 +130,7 @@
     h2 {
       text-align: center;
     }
-    
+
     p.lead {
       text-align: center;
     }
