+++ src/assets/css/sass/home/_searchResults.scss
@@ -16,7 +16,7 @@
       margin: 10px 0 0 15px;
       padding: 0;
     }
-    
+
     p {
       font-size: 20px;
       margin-bottom: 0;
