+++ src/assets/css/sass/registration/_registration-page.scss
@@ -47,7 +47,7 @@
     padding: .5em 2em !important;
     border-radius: 8px;
   }
-  
+
   .btn:hover {
     border-color: $gray !important;
     box-shadow: none !important;
