+++ src/languages/en.json
@@ -1,5 +1,5 @@
 {
-  "home_title": "rif name service",
+  "home_title": "RIF name service",
   "home_subtitle": "get your domain, forget your address",
   "register_your_domain": "register your domain",
   "what_is_rns_title": "what is rns?",
