+++ src/languages/es_uy.json
@@ -1,5 +1,5 @@
 {
-  "home_title": "rif name service",
+  "home_title": "RIF name service",
   "home_subtitle": "obtenga su dominio, olvide su dirección",
   "register_your_domain": "registre su dominio",
   "what_is_rns_title": "¿qué es rns?",
