+++ src/languages/ja.json
@@ -1,5 +1,5 @@
 {
-  "home_title": "rif name service",
+  "home_title": "RIF name service",
   "home_subtitle": "ドメインを取得する、アドレスを紛失する",
   "register_your_domain": "ドメインを登録する",
   "what_is_rns_title": "rns とは?",
