+++ src/languages/pt_br.json
@@ -1,5 +1,5 @@
 {
-  "home_title": "rif name service",
+  "home_title": "RIF name service",
   "home_subtitle": "obtenha seu domínio, esqueça seu endereço",
   "register_your_domain": "registre seu domínio",
   "what_is_rns_title": "o que é o rns?",
